Add the template create to multi-room editor.
