+++ src/modlunky2/ui/levels/vanilla_levels/levels_tree.py
@@ -15,6 +15,9 @@ from modlunky2.levels.level_templates import (
     TemplateSetting,
 )
 from modlunky2.ui.widgets import PopupWindow
+from modlunky2.ui.levels.shared.create_template_dialog import (
+    present_create_template_dialog,
+)
 from modlunky2.ui.levels.shared.tags import TAGS
 from modlunky2.ui.levels.shared.setrooms import Setroom, MatchedSetroom
 
@@ -87,170 +90,29 @@ class LevelsTree(ttk.Treeview):
         if parent_iid:
             self.on_select_room()
         elif self.index(item_iid) == len(self.rooms):
-            self.create_template_dialog()
+            present_create_template_dialog(self.config, self.create_template)
 
-    def create_template_dialog(self):
-        win = PopupWindow("Create Template", self.config)
-
-        row = 0
-        values_frame = tk.Frame(win)
-        values_frame.grid(row=row, column=0, sticky="nw")
-        row = row + 1
-
-        values_row = 0
-        name_lbl = ttk.Label(values_frame, text="Name: ")
-        name_ent = ttk.Entry(values_frame)
-        # name_ent.insert(0, "New_Pack")  # Default to rooms current name
-        name_lbl.grid(row=values_row, column=0, pady=2, sticky="e")
-        name_ent.grid(row=values_row, column=1, pady=2, sticky="we")
-
-        values_row += 1
-
-        comment_lbl = ttk.Label(values_frame, text="Comment: ")
-        comment_ent = ttk.Entry(values_frame)
-        # name_ent.insert(0, "New_Pack")  # Default to rooms current name
-        comment_lbl.grid(row=values_row, column=0, pady=2, sticky="e")
-        comment_ent.grid(row=values_row, column=1, pady=2, sticky="we")
-
-        values_row += 1
-
-        size_type_label = tk.Label(values_frame, text="Size in: ")
-        size_type_label.grid(row=values_row, column=0, sticky="e", pady=2)
-
-        size_type_combobox = ttk.Combobox(values_frame, height=25)
-        size_type_combobox.set("Subrooms")
-        size_type_combobox.grid(row=values_row, column=1, sticky="we", pady=2)
-        size_type_combobox["state"] = "readonly"
-        size_type_combobox["values"] = ["Subrooms", "Tiles"]
-        size_type_combobox.current(0)
-
-        values_row += 1
-
-        width_label = tk.Label(values_frame, text="Width: ")
-        width_label.grid(row=values_row, column=0, sticky="e", pady=2)
-
-        width_ent = ttk.Entry(values_frame)
-        width_ent.insert(0, "1")
-        width_ent.grid(row=values_row, column=1, sticky="we", pady=2)
-
-        values_row += 1
-
-        height_label = tk.Label(values_frame, text="Height: ")
-        height_label.grid(row=values_row, column=0, sticky="e", pady=2)
-
-        height_ent = ttk.Entry(values_frame)
-        height_ent.insert(0, "1")
-        height_ent.grid(row=values_row, column=1, sticky="we", pady=2)
-
-        values_row += 1
-
-        error_lbl = tk.Label(values_frame, text="", fg="red")
-        error_lbl.grid(row=values_row, column=0, columnspan=2)
-        error_lbl.grid_remove()
-
-        values_row += 1
-
-        def update_then_destroy_pack():
-            template_name = ""
-            for char in str(name_ent.get()):
-                if str(char) != " ":
-                    template_name += str(char)
-                else:
-                    template_name += "_"
-
-            comment = comment_ent.get()
-            if len(comment) == 0:
-                comment = None
-
-            width_str = width_ent.get()
-            height_str = height_ent.get()
-            if not width_str.isdecimal():
-                error_lbl.grid()
-                error_lbl["text"] = "Invalid width."
-                return
-
-            if not height_str.isdecimal():
-                error_lbl.grid()
-                error_lbl["text"] = "Invalid height."
-                return
-
-            width = int(width_str)
-            height = int(height_str)
-            if size_type_combobox.current() != 1:
-                width *= 10
-                height *= 8
-
-            error_lbl.grid_remove()
-            if width == 0:
-                error_lbl.grid()
-                error_lbl["text"] = "Invalid width."
-                return
-
-            if height == 0:
-                error_lbl.grid()
-                error_lbl["text"] = "Invalid height."
-                return
-
-            if len(template_name) == 0:
-                error_lbl.grid()
-                error_lbl["text"] = "Name the template."
-                return
-
-            success, error_message, new_room = self.on_create_template(
-                template_name, comment, width, height
-            )
-            if success:
-                # self.delete(len(self.rooms))
-
-                room_display_name = str(new_room.name)
-                if new_room.comment is not None and len(str(new_room.comment)) > 0:
-                    room_display_name += "   // " + new_room.comment
-                new_template_item = self.insert(
-                    "", len(self.rooms), text=room_display_name
+    def create_template(self, name, comment, width, height):
+        success, error_message, new_room = self.on_create_template(
+            name, comment, width, height
+        )
+        if success:
+            room_display_name = str(new_room.name)
+            if new_room.comment is not None and len(str(new_room.comment)) > 0:
+                room_display_name += "   // " + new_room.comment
+            new_template_item = self.insert("", len(self.rooms), text=room_display_name)
+            for room in new_room.rooms:
+                child_id = self.insert(
+                    new_template_item, "end", text=room.name or "room"
                 )
-                for room in new_room.rooms:
-                    child_id = self.insert(
-                        new_template_item, "end", text=room.name or "room"
-                    )
-                    self.focus(child_id)
-                    self.selection_set(child_id)
+                self.focus(child_id)
+                self.selection_set(child_id)
 
-                self.item(new_template_item, open=True)
-                self.rooms.append(new_room)
-                self.on_select_room()
-                win.destroy()
-            else:
-                error_lbl.grid()
-                error_lbl["text"] = error_message or "Error creating template."
-                return
-            # name_ent.delete(0, "end")
-            # name_ent.insert(0, pack_name)
-            # if not os.path.isdir(self.packs_path / str(name_ent.get())):
-            #     os.mkdir(self.packs_path / str(name_ent.get()))
-            #     self.load_packs()
-            #     win.destroy()
-            # else:
-            #     logger.warning("Pack name taken")
-            #     name_ent.delete(0, "end")
-            #     name_ent.insert(0, "Name Taken")
+            self.item(new_template_item, open=True)
+            self.rooms.append(new_room)
+            self.on_select_room()
 
-        separator = ttk.Separator(win)
-        separator.grid(row=row, column=0, columnspan=2, pady=5, sticky="news")
-
-        row = row + 1
-
-        buttons = ttk.Frame(win)
-        buttons.grid(row=row, column=0, sticky="news")
-        buttons.columnconfigure(0, weight=1)
-        buttons.columnconfigure(1, weight=1)
-
-        ok_button = ttk.Button(buttons, text="Create", command=update_then_destroy_pack)
-        ok_button.grid(row=0, column=0, pady=5, sticky="news")
-
-        cancel_button = ttk.Button(buttons, text="Cancel", command=win.destroy)
-        cancel_button.grid(row=0, column=1, pady=5, sticky="news")
-
-        row = row + 1
+        return success, error_message
 
     def popup(self, event):
         try:
